lots of experiments and fixes

+++ ppmtolab.cc
@@ -14,10 +14,10 @@ using namespace makemore;
 int main(int argc, char **argv) {
   PPM p;
   p.read(stdin);
-  std::vector<double> v;
+  std::vector<uint8_t> v;
   p.vectorize(&v);
   assert(v.size() == p.w * p.h * 3);
-  int ret = fwrite(v.data(), sizeof(double), v.size(), stdout);
+  int ret = fwrite(v.data(), sizeof(uint8_t), v.size(), stdout);
   assert(ret == p.w * p.h * 3);
   return 0;
 }
